move remove signup to roster service
